+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+**1.1.0** (2024-30-02)
+- Support new version of the forger contract released in eon 1.3.0
+- Support for new methods of the smart contract: stakeOf, getPagedForgersStakes, getPagedForgersStakesByUser
+
 **1.0.0** (2024-07-02)
 - Initial release